timshee/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
Deprecated==1.2.14
distro==1.9.0
Django==5.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-mjml==1.3
django-oauth-toolkit==2.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-social-oauth2==2.1.3
ecdsa==0.18.0
filelock==3.15.4
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.6
magic-filter==1.0.12
Markdown==3.5.2
multidict==6.0.5
netaddr==1.2.1
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
setuptools==69.1.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
yookassa==3.1.0
